packages/qa-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qa-agent"
version = "0.1.0"
description = "Game knowledge Q&A agent for Sanmou."
requires-python = ">=3.11"
dependencies = [
  "sanmou-common",
  "pydantic>=2.6,<3",
  "PyYAML>=6.0,<7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
